Resolve circular dependency by splitting createSearchIndex function out into its own file
